requirements.txt
# Core FastAPI Dependencies
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.5
python-multipart==0.0.20
httpx==0.27.0

# Database Dependencies
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# Authentication Dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI/ML Dependencies
openai>=1.81.0

# RAG Pipeline Dependencies (required for existing RAG service)
llama-index==0.12.5
llama-index-embeddings-openai==0.3.1
llama-index-vector-stores-qdrant==0.8.0
qdrant-client==1.12.1
unstructured[pdf]==0.18.14
tiktoken==0.8.0
nest-asyncio==1.6.0
